# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book해도. REST API",
    "version": "2.0.0",
    "description": "쿠키 인증, 소유권 검증, 날짜별 optimistic concurrency. 전체 일정/메모/대안 변경에 expectedRevision 필수(누락 428, 충돌 409). POST 장소 추가는 원자적 append. X-Request-ID로 오류 추적."
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "동일 출처 API"
    }
  ],
  "paths": {
...
    "/trips/{id}/days": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "summary": "여행에 날짜 추가",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "최대 일수 초과 / 입력 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "여행 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "이미 존재하는 날짜",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "여행"
        ],
        "operationId": "post__trips_id_days"
      }
    },
...
    "/trips/{id}/days/{date}/day-alternatives": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "date",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "date"
          }
        }
      ],
      "get": {
        "summary": "날씨와 동선 기반 하루 전체 코스 추천·미리보기",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayAlternativeResult"
                }
              }
            }
          },
          "400": {
            "description": "입력 형식 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "본인 여행·날짜 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "코스 변경",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AUTO",
                "INDOOR",
                "NEARBY"
              ]
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 3,
              "maximum": 6,
              "default": 4
            }
          }
        ],
        "description": "날씨가 좋거나 예보가 없어도 사용할 수 있다. 현재 일정 중심의 직선 20km 후보에서 날씨 맞춤, 실내 중심, 가까운 곳 중심 코스를 생성한다. strategy를 지정하면 선택 코스의 인접 실제 경로를 조회하지만 저장하지 않는다. 장소 운영 여부와 최적 동선을 보장하지 않는다.",
        "tags": [
          "추천"
        ],
        "operationId": "get__trips_id_days_date_day_alternatives"
      },
      "patch": {
        "summary": "사용자 확정 후 하루 전체 일정을 추천 코스로 교체",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saved": {
                      "type": "boolean"
                    },
                    "revision": {
                      "type": "integer",
                      "minimum": 0,
                      "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "장소 수·거리·입력 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "본인 여행·날짜 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "조회 이후 일정·메모가 변경됨. 최신 내용 확인 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "428": {
            "description": "expectedRevision 누락",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "placeIds",
                  "expectedPlaceIds",
                  "expectedRevision"
                ],
                "properties": {
                  "placeIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 2,
                    "maxItems": 6,
                    "uniqueItems": true
                  },
                  "expectedPlaceIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 30,
                    "uniqueItems": true,
                    "description": "미리보기 시점의 기존 하루 일정 순서"
                  },
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
                  }
                }
              }
            }
          }
        },
        "description": "날짜 행을 잠그고 기존 일정 순서와 expectedPlaceIds를 비교한다. 추천 코스의 장소 존재·중복·기존 중심과의 거리를 검증하고 트랜잭션으로 전체 순서를 교체한다. 새 코스에 남는 장소의 메모만 유지하며 오류나 충돌 시 기존 일정을 보존한다. expectedRevision을 잠긴 날짜의 현재 revision과 비교하며 성공하면 revision을 증가시킨다.",
        "tags": [
          "추천"
        ],
        "operationId": "patch__trips_id_days_date_day_alternatives"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kita_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "currentRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "regionId": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "ATTRACTION",
              "RESTAURANT",
              "LODGING"
            ]
          },
          "name": {
            "type": "string"
          },
          "nameJa": {
            "type": "string"
          },
          "nameKo": {
            "type": [
              "string",
              "null"
            ]
          },
          "nameEn": {
            "type": [
              "string",
              "null"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "municipality": {
            "type": [
              "string",
              "null"
            ]
          },
          "website": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "openingHours": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "source": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "license": {
                  "type": "string"
                }
              }
            }
          },
          "operatingPeriod": {
            "type": "null"
          },
          "seasonNotice": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "trendBadge": {
            "type": [
              "string",
              "null"
            ]
          },
          "reviewSummary": {
            "type": "object",
            "properties": {
              "rating": {
                "type": "number"
              },
              "reviewCount": {
                "type": "integer"
              },
              "japaneseCount": {
                "type": "integer"
              },
              "sampleCount": {
                "type": "integer"
              },
              "score": {
                "type": "number"
              }
            }
          },
          "recommended": {
            "type": "boolean"
          },
          "recommendationReasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedCost": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 100000000,
            "description": "사용자가 입력한 장소 예상 총액 (JPY). null은 미입력."
          },
          "startMinute": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 1439
          },
          "endMinute": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 1440
          }
        }
      },
...
      "Segment": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "uuid"
          },
          "to": {
            "type": "string",
            "format": "uuid"
          },
          "distanceMeters": {
            "type": "number"
          },
          "durationSeconds": {
            "type": [
              "number",
              "null"
            ]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          "estimatedCost": {
            "type": "null"
          },
          "costNotice": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "transitFare": {
            "type": [
              "string",
              "object",
              "null"
            ]
          },
          "tolls": {
            "type": [
              "object",
              "null"
            ]
          },
          "costs": {
            "type": [
              "object",
              "null"
            ]
          },
          "polyline": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "valhalla",
              "google",
              "straight-line"
            ]
          },
          "notice": {
            "type": "string"
          }
        }
      },
      "Weather": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "high": {
            "type": "number"
          },
          "low": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "notice": {
            "type": "string"
          },
          "precipitationMm": {
            "type": [
              "number",
              "null"
            ]
          },
          "snowfallCm": {
            "type": [
              "number",
              "null"
            ]
          },
          "precipitationProbability": {
            "type": [
              "number",
              "null"
            ]
          },
          "windSpeedKmh": {
            "type": [
              "number",
              "null"
            ],
            "description": "일 최고 10m 풍속 km/h"
          }
        }
      },
...
      "DayAlternativeResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "NEEDS_ANCHOR",
              "NO_CANDIDATES"
            ]
          },
          "weather": {
            "$ref": "#/components/schemas/Weather"
          },
          "weatherMode": {
            "type": "string",
            "enum": [
              "ADVERSE",
              "FAIR",
              "UNKNOWN"
            ]
          },
          "expectedPlaceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "plans": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "enum": [
                    "AUTO",
                    "INDOOR",
                    "NEARBY"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                },
                "places": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                },
                "distanceMeters": {
                  "type": "number"
                }
              }
            }
          },
          "preview": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "properties": {
                  "plan": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "enum": [
                          "AUTO",
                          "INDOOR",
                          "NEARBY"
                        ]
                      },
                      "label": {
                        "type": "string"
                      },
                      "reason": {
                        "type": "string"
                      },
                      "places": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Place"
                        }
                      },
                      "distanceMeters": {
                        "type": "number"
                      }
                    }
                  },
                  "segments": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Segment"
                    }
                  },
                  "complete": {
                    "type": "boolean"
                  },
                  "distanceMeters": {
                    "type": [
                      "number",
                      "null"
                    ]
                  },
                  "durationSeconds": {
                    "type": [
                      "number",
                      "null"
                    ]
                  }
                }
              }
            ]
          },
          "notice": {
            "type": "string"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
          }
        }
      },
...
    }
  }
}